+++ src/main/java/com/real/realoasis/domain/couple/domain/entity/Couple.java
@@ -1,6 +1,7 @@
 package com.real.realoasis.domain.couple.domain.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.real.realoasis.domain.heart.domain.entity.Heart;
 import com.real.realoasis.domain.user.domain.entity.User;
 import com.real.realoasis.global.entity.BaseIdEntity;
 import lombok.*;
@@ -8,6 +9,8 @@ import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
 import javax.persistence.*;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 
 @Entity
 @Getter
@@ -16,15 +19,47 @@ import javax.persistence.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Couple extends BaseIdEntity {
     @Column(nullable = false)
-    private String coupleId;
+    private String startDay;
     @Column(nullable = false)
-    private String code;
-    @JsonIgnore
-    @ManyToOne(fetch = FetchType.LAZY)
-    @OnDelete(action = OnDeleteAction.CASCADE)
-    private User user;
+    private String today;
+    @Column(nullable = false)
+    private long datingDate;
+    @Column(nullable = false)
+    private long anniversaryDate;
 
-    public void update(User coupleUser) {
-        this.coupleId = coupleUser.getId();
+    @JsonIgnore
+    @OneToOne(fetch = FetchType.LAZY)
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    private User userA;
+
+    @JsonIgnore
+    @OneToOne(fetch = FetchType.LAZY)
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    private User userB;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    private Heart heart;
+
+
+    public void updateStartDay(String startDay){
+        this.startDay = startDay;
+    }
+
+    @PrePersist
+    public void today(){
+        this.today = LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
+    }
+
+    public void updateDatingDate(long datingDate) {
+        this.datingDate = datingDate;
+    }
+
+    public void updateAnniversaryTime(long anniversaryDate) {
+        this.anniversaryDate = anniversaryDate;
+    }
+
+    public void updateHeart(Heart heart) {
+        this.heart = heart;
     }
 }
